api/requirements.txt
langchain
langchain-core
langchain_community
langchain-huggingface
langchain-google-genai
mistralai
docx2txt
pypdf
langchain_chroma
python-multipart
python-dotenv
streamlit
fastapi
google-generativeai
uvicorn

opencv-python
pdf2image
PyMuPDF
Pillow
requests


# Gradio UI Requirements
gradio>=4.0.0
typing-extensions